+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 * Prevent nil panic by checking for an error when creating a new request
 
+* Hide OAuth2 authorization cookie value and Xsrf Token when hide sensitive information is enabled
+
 ## Released
 
 ### v0.8.0